Fix showing new pieces which match the current filter after an update
